succesfully added permission

+++ resources/views/adminview/index.blade.php
@@ -30,7 +30,10 @@
 
                         </div>
                         <div class="pull-right">
-                            <a class="btn btn-success" href="{{ route('users.create') }}"> Create New Admin</a>
+                            @can('admin-create')
+                                <a class="btn btn-success" href="{{ route('users.create') }}"> Create New Admin</a>
+                            @endcan
+
                         </div>
                     </div>
                 </div>
@@ -65,10 +68,16 @@
                             </td>
                             <td>
                                 <a class="btn btn-info" href="{{ route('users.show', $user->id) }}">Show</a>
-                                <a class="btn btn-primary" href="{{ route('users.edit', $user->id) }}">Edit</a>
-                                {!! Form::open(['method' => 'DELETE', 'route' => ['users.destroy', $user->id], 'style' => 'display:inline']) !!}
-                                {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
-                                {!! Form::close() !!}
+                                @can('admin-edit')
+                                    <a class="btn btn-primary" href="{{ route('users.edit', $user->id) }}">Edit</a>
+                                @endcan
+                                @can('admin-delete')
+                                    {!! Form::open(['method' => 'DELETE', 'route' => ['users.destroy', $user->id], 'style' => 'display:inline']) !!}
+                                    {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
+                                    {!! Form::close() !!}
+                                @endcan
+
+
                             </td>
                         </tr>
                     @endforeach
